requirements.txt
astroid==2.3.3
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.8.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.1
entrypoints==0.3
epc==0.0.5
flake8==3.7.9
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
importmagic==0.1.7
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==7.2.0
mypy==0.740
mypy-extensions==0.4.3
oauthlib==3.1.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.4.4
pyls==0.1.6
pyls-isort==0.1.1
pyls-mypy==0.1.8
python-jsonrpc-server==0.3.0
python-language-server==0.31.1
PyYAML==5.1.2
RarbgAPI==0.3.1
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.0
sexpdata==0.0.3
six==1.13.0
snowballstemmer==2.0.0
soupsieve==1.9.5
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.7.4.1
ujson==1.35
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
wrapt==1.11.2
yapf==0.28.0
zipp==0.6.0
